Добавил папку CrossSectionDimension внутрь нее убрал класс Section manager
